Move HN API code into lib folder
