+++ .gitignore
@@ -17,3 +17,5 @@
 /yarn-error.log
 
 .byebug_history
+lib/assets/.keep
+lib/tasks/.keep
