+++ spec/lib/hacker_news_api_spec.rb
